Make build size of JSX runtime smaller
by splitting internal methods for JSX into another file
