+++ src/block-kit/elements/ChannelsSelect.ts
@@ -3,7 +3,7 @@ import {
   MultiChannelsSelect,
   InputBlock,
 } from '@slack/types'
-import { BuiltInComponent, createComponent } from '../../jsx'
+import { BuiltInComponent, createComponent } from '../../jsx-internals'
 import { DistributedProps, coerceToInteger } from '../../utils'
 import { ConfirmableProps } from '../composition/Confirm'
 import { plainText } from '../composition/utils'
