+++ src/block-kit/elements/RadioButtonGroup.ts
@@ -1,6 +1,7 @@
 import { RadioButtons as RadioButtonsElement, InputBlock } from '@slack/types'
 import { JSXSlackError } from '../../error'
-import { JSXSlack, BuiltInComponent, createComponent } from '../../jsx'
+import { JSXSlack } from '../../jsx'
+import { BuiltInComponent, createComponent } from '../../jsx-internals'
 import { ConfirmableProps } from '../composition/Confirm'
 import {
   RadioButton,
