+++ src/block-kit/layout/Divider.ts
@@ -1,5 +1,5 @@
 import { DividerBlock } from '@slack/types'
-import { createComponent } from '../../jsx'
+import { createComponent } from '../../jsx-internals'
 import { LayoutBlockProps } from './utils'
 
 export interface DividerProps extends LayoutBlockProps {
