+++ src/block-kit/layout/File.ts
@@ -1,5 +1,5 @@
 import { FileBlock } from '@slack/types'
-import { createComponent } from '../../jsx'
+import { createComponent } from '../../jsx-internals'
 import { LayoutBlockProps } from './utils'
 
 export interface FileProps extends LayoutBlockProps {
